Update albumController.js
The `async.parallel()` method is passed an object with functions for getting the counts for each models. These functions are all started at the same time. When all them have completed the final callback is invoked with the counts in the results parameter (or an error).

on success the callback function calls [res.render()](http://expressjs.com/en/4x/api.html#res.render), specifying a view(template) named 'index' and and object containing the data that is to be inserted into it (this includes the results objects that contains our model counts). The data is supplied as key-value pairs, and can be accessed in the template using the key.