Update index.pug
The view is straightforward. We extend the layout.pug base template, overriding the block named ‘content’. The first h1 heading will be the escaped text for the title variable that was passed into the render() function-note the use of the ‘h1=’ so that the following text is treated as a JavaScript expression. We then include a paragraph introducing the MusicLibrary.