Update albumController.js
The method uses the models find() function to return all Album objects, selecting to return only the title and artist as we don't need the other fields (it will also return the _id and virtual fields). Here we also call populate() on Album, specifying the artist field-this will replace the stored album artist id with the full artist detail.